+++ docs/releases.md
@@ -1,3 +1,9 @@
+# Unreleased
+- Allow commenting on first station (thanks Andrew Atkinson)
+- Therion export improvements (thanks Andrew Atkinson)
+- Setting for turning on/off floating buttons on Table view (thanks Andrew Atkinson)
+- Added "Add Leg" and "Add Splay" to the Tools menu
+
 # 2026-04-28 1.11.2
 - Fix a bug causing cross-sections to be initially hidden